Add files via upload
Still working on it, but I like the layout here a lot more than in the first on. Feels neater. That being said the mobil version is a nightmare to look at right now. I'm going to see how I can fix that up. The links on the side bar will be gone and that's where I plan to put my socials, email and skills. I still want the moving bars but I'm going to keep them simple.

The background around the boxes won't be there in the final. It's just so I can lay them out right.